moved config file and updated requires
